Cleanup of specfile & patching to reduce warnings
There was a bunch of things in the specfile that were not needed (yet).
Holding on to them so that when we're ready to build with our own
systemd version that we can fire away.

Also added a patch (submitted as PR# 3665 to rkt) to lower the needless
warnings from unsquashfs.