bug/yum: dist macro used in place of $releasever
Previously when building the RPM the dist macro was used in place of the
host variable $releasever.  This leads to a problem because dist embeds
a dot in the version as well as results in a hard coding of the path
regardless of the version of the underlying operating system.