Drop unused files from the repository
Since none of these were actually being used, drop them to avoid
misleading anyone that might be reading this.  Some of this could
be nice and might be worth resurrecting, but that can wait until
there is a user for it.

The documentation has been dropped because the official website is
pulling a newer version from the Tectonic Installer repository.

The signing directory has been dropped since after its original use
was abandoned, it only contained files that are duplicates with RPM
source files.

The scripts directory only had a script for signing, which has been
obsoleted (as above).

The user-data directory was never used.  Building and testing is
currently happening as a mix of Jenkins jobs and local commands.

The Tito + Makefile build setup was dropped since it was incomplete
and did not produce correct packages.  For the time being, it is
replaced by just running Mock with mock.cfg on Fedora 26 to build
correctly.

Since every other directory was dropped, the contents of the rpm
directory was just moved to the main directory.